requirements.txt
langchain>=0.3.0
langchain-core>=0.3.0
langchain-google-genai>=2.0.0
langgraph>=0.2.0
fastapi>=0.115.0
uvicorn>=0.30.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
chromadb>=0.5.0
python-dotenv>=1.0.0
pydantic>=2.0.0
aiosmtplib>=3.0.0
jinja2>=3.1.0
python-multipart>=0.0.9
google-genai>=1.0.0
websockets>=12.0